Add useMessageOpening hook
